Move MarkdownFileParser into Markdown namespace
