+++ packages/framework/src/Models/Parsers/MarkdownPageParser.php
@@ -5,7 +5,7 @@ namespace Hyde\Framework\Models\Parsers;
 use Hyde\Framework\Contracts\AbstractPageParser;
 use Hyde\Framework\Hyde;
 use Hyde\Framework\Models\Pages\MarkdownPage;
-use Hyde\Framework\Services\MarkdownFileParser;
+use Hyde\Framework\Modules\Markdown\MarkdownFileParser;
 
 /**
  * Parses a Markdown file into a MarkdownPage object using the MarkdownPage intermediary.
